Bugfix: Ensure jobs are scheduled on startup
An important feature went missing during the refactor: the scheduling of jobs at the start of the service. This feature has now been restored.

Note: Since the flow has changed, jobs may now have their own schedules and can be programmed in the database. We need to wait for the database to come up before starting the scheduling. This contrasts with the previous flow, where jobs were controlled by a single programmatically defined cron pattern.